Addresses three trivial compiler warnings in code. Caught these while checking the #4132 DialogCode patch.
git-svn-id: http://svn.mythtv.org/svn/trunk@14856 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythbackend/backendutil.cpp
@@ -179,7 +179,7 @@ void BackendQueryDiskSpace(QStringList &strlist,
             // Sometimes the space reported for an NFS mounted dir is slightly
             // different than when it is locally mounted because of block sizes
             if ((absLongLong(it1->totalSpaceKB - it2->totalSpaceKB) <= 16 ) &&
-                (absLongLong(it1->usedSpaceKB - it2->usedSpaceKB)
+                ((size_t)absLongLong(it1->usedSpaceKB - it2->usedSpaceKB)
                  < maxWriteFiveSec))
             {
                 if (!it1->hostname.contains(it2->hostname))
@@ -261,7 +261,7 @@ void GetFilesystemInfos(QMap<int, EncoderLink*> *tvList,
             // different than when it is locally mounted because of block sizes
             if ((it2->fsID == -1) &&
                 ((absLongLong(it1->totalSpaceKB - it2->totalSpaceKB) <= 16 ) &&
-                 (absLongLong(it1->usedSpaceKB - it2->usedSpaceKB)
+                 ((size_t)absLongLong(it1->usedSpaceKB - it2->usedSpaceKB)
                   < maxWriteFiveSec)))
                 it2->fsID = it1->fsID;
         }
